splitted DateTimeInput to DateInput and DateTimeInput, removed DateTimeInput::$defaultDateTimeFormat and DateTimeInput::$defaultDateFormat and instead introduced $defaultFormat on DateTimeInput and DateInput, introduced $additionalHtmlClasses on DateInput  and DateTimeInput so it's easier to use with datepickers
